// src/enums/player/PlayerStats.ts
/* eslint-disable @typescript-eslint/no-duplicate-enum-values */
export enum PlayerStats {
  KazumaHp = 40,
  KazumaAttackMin = 6,
  KazumaAttackMax = 12,

  DarknessHp = 80,
  DarknessAttackMin = 0,
  DarknessAttackMax = 6,

  MeguminHp = 30,
  MeguminAttackMin = 25,
  MeguminAttackMax = 25,

  AquaHp = 40,
  AquaAttackMin = 6,
  AquaAttackMax = 12,
}