// src/enums/player/PlayerThmb.ts
export enum PlayerThmb {
  Kazuma = 'https://fox3000foxy.com/konosuba-rpg/assets/player/thmb_in_1001100.png',
  Darkness = 'https://fox3000foxy.com/konosuba-rpg/assets/player/thmb_in_1031100.png',
  Megumin = 'https://fox3000foxy.com/konosuba-rpg/assets/player/thmb_in_1021100.png',
  Aqua = 'https://fox3000foxy.com/konosuba-rpg/assets/player/thmb_in_1011100.png',
}